refactor: improve modularity of read operations
